Got it to build but ... operator delete is nasty

+++ include/nul/service_config.h
@@ -42,10 +42,10 @@ struct ConfigProtocol : public GenericProtocol {
     {}
   };
 
-  unsigned start_config(Utcb &utcb, unsigned short &id, unsigned long &maxmem,
+  unsigned long start_config(Utcb &utcb, unsigned short &id, unsigned long &maxmem,
                         cap_sel cap_sc_usage, char const * config, unsigned long len = ~0UL)
   {
-    unsigned res;
+    unsigned long res;
     Utcb & utcb2 = init_frame(utcb, TYPE_START_CONFIG) << Utcb::String(config, len) << Crd(cap_sc_usage, 0, DESC_CAP_ALL);
     if (!utcb2.validate_send_bounds()) {res = ECONFIGTOOBIG; goto out; }
     res = call_server(utcb2, false);
